tsx를 ts로 변경
useState왜 오류나지
