+++ .vscode/settings.json
@@ -3,7 +3,7 @@
       80
   ],
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
+    "source.fixAll.eslint": "explicit"
   },
   "trailing-spaces.trimOnSave": true
 }
