+++ src/App.test.tsx
@@ -1,9 +1,10 @@
 import { render, screen } from '@testing-library/react';
+import React from 'react';
 
-import App from './App';
+import App2 from './App';
 
 test('App', () => {
-  render(<App />);
+  render(React.createElement(App2, null));
 
   screen.getByText(/Hello, wholeman!/);
 });
