+++ src/components/Greeting.ts
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Greeting(name: string) {
+  return React.createElement('p', null, 'Hello,', ' ', name, '!');
+}
