+++ src/main.ts
@@ -1,6 +1,7 @@
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './App';
+import App2 from './App';
 
 function main() {
   const container = document.getElementById('root');
@@ -9,7 +10,7 @@ function main() {
   }
 
   const root = ReactDOM.createRoot(container);
-  root.render(<App />);
+  root.render(App2());
 }
 
 main();
